Bootstrap the likelihood-ratio test statistic for mixture components

Bootstrap the likelihood-ratio test statistic for mixture components

bootstrap_lrt(...)

Arguments

df
data.frame with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles
the number of profiles (or mixture components) to be estimated

variances
how the variable variances are estimated; defaults to "equal" (to be constant across profiles); other option is "varying" (to be varying across profiles)

covariances
how the variable covariances are estimated; defaults to "zero" (to not be estimated, i.e. for the covariance matrix to be diagonal); other options are "varying" (to be varying across profiles) and "equal" (to be constant across profiles)

model
which model to estimate (DEPRECATED; use variances and covariances instead)

center_raw_data
logical for whether to center (M = 1) the raw data (before clustering); defaults to FALSE

scale_raw_data
logical for whether to scale (SD = 1) the raw data (before clustering); defaults to FALSE

to_return
character string for either "tibble" (or "data.frame") or "mclust" if "tibble" is selected, then data with a column for profiles is returned; if "mclust" is selected, then output of class mclust is returned

return_posterior_probs
TRUE or FALSE (only applicable if to_return == "tibble"); whether to include posterior probabilities in addition to the posterior profile classification; defaults to TRUE

return_orig_df
TRUE or FALSE (if TRUE, then the entire data.frame is returned; if FALSE, then only the variables used in the model are returned)

prior_control
whether to include a regularizing prior; defaults to false

print_which_stats
if set to "some", prints (as a message) the log-likelihood, BIC, and entropy; if set to "all", prints (as a message) all information criteria and other statistics about the model; if set to any other values, then nothing is printed

Details

Bootstrap the p-values for the likelihood-ratio test statistic for the number of mixture components for an mclust model.

See also