Explore BIC of MPlus models

Explore BIC of MPlus models

compare_models_mplus(df, ..., n_profiles = 2:10, models = 1:5,
  start_iterations = c(100, 20), m_iterations = 500, save_models = NULL)

Arguments

df
data.frame with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles_range
a vector with the range of the number of mixture components to explore; defaults to 2 through 10 (2:10)

Value

a list with a data.frame with the BIC values and a list with all of the model output; if save_models is the name of an rds file (i.e., "out.rds"), then the model output will be written with that filename and only the data.frame will be returned

Details

Explore the BIC values of a range of Mplus models in terms of a) the structure of the residual covariance matrix and b) the number of mixture components (or profiles)

Examples

compare_models_mplus(iris, Sepal.Length, Sepal.Width, Petal.Length, Petal.Width)

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 0/54

#> Processed model with n_profiles = 2 and model = 1

#> Result: BIC = 1042.968

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 1/54

#> Processed model with n_profiles = 2 and model = 2

#> Result: BIC = 688.097

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 2/54

#> Processed model with n_profiles = 2 and model = 3

#> Result: BIC = 857.551

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 3/54

#> Processed model with n_profiles = 2 and model = 4

#> Result: BIC = 672.333

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 4/54

#> Processed model with n_profiles = 2 and model = 5

#> Result: BIC = 574.018

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 5/54

#> Processed model with n_profiles = 3 and model = 1

#> Result: BIC = 813.042

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 6/54

#> Processed model with n_profiles = 3 and model = 2

#> Result: BIC = 632.963

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 7/54

#> Processed model with n_profiles = 3 and model = 3

#> Result: BIC = 743.997

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 8/54

#> Processed model with n_profiles = 3 and model = 4

#> Result: BIC = 618.349

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 9/54

#> Processed model with n_profiles = 3 and model = 5

#> Result: BIC = 580.839

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 10/54

#> Processed model with n_profiles = 4 and model = 1

#> Result: BIC = 735.479

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 11/54

#> Processed model with n_profiles = 4 and model = 2

#> Result: BIC = 591.406

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 12/54

#> Processed model with n_profiles = 4 and model = 3

#> Result: BIC = 705.067

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 13/54

#> Processed model with n_profiles = 4 and model = 4

#> Result: BIC = 607.269

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 14/54

#> Processed model with n_profiles = 4 and model = 5

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 15/54

#> Processed model with n_profiles = 5 and model = 1

#> Result: BIC = 694.389

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 16/54

#> Processed model with n_profiles = 5 and model = 2

#> Result: BIC = 595.889

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 17/54

#> Processed model with n_profiles = 5 and model = 3

#> Result: BIC = 700.902

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 18/54

#> Processed model with n_profiles = 5 and model = 4

#> Result: BIC = 638.968

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 19/54

#> Processed model with n_profiles = 5 and model = 5

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 20/54

#> Processed model with n_profiles = 6 and model = 1

#> Result: BIC = 672.595

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 21/54

#> Processed model with n_profiles = 6 and model = 2

#> Result: BIC = 598.972

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 22/54

#> Processed model with n_profiles = 6 and model = 3

#> Result: BIC = 704.198

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 23/54

#> Processed model with n_profiles = 6 and model = 4

#> Result: BIC = 638.26

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 24/54

#> Processed model with n_profiles = 6 and model = 5

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 25/54

#> Processed model with n_profiles = 7 and model = 1

#> Result: BIC = 669.482

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 26/54

#> Processed model with n_profiles = 7 and model = 2

#> Result: BIC = 610.265

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 27/54

#> Processed model with n_profiles = 7 and model = 3

#> Result: BIC = 704.525

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 28/54

#> Processed model with n_profiles = 7 and model = 4

#> Result: BIC = 680.795

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 29/54

#> Processed model with n_profiles = 7 and model = 5

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 30/54

#> Processed model with n_profiles = 8 and model = 1

#> Result: BIC = 661.074

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 31/54

#> Processed model with n_profiles = 8 and model = 2

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 32/54

#> Processed model with n_profiles = 8 and model = 3

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 33/54

#> Processed model with n_profiles = 8 and model = 4

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 34/54

#> Processed model with n_profiles = 8 and model = 5

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 35/54

#> Processed model with n_profiles = 9 and model = 1

#> Result: BIC = 661.624

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 36/54

#> Processed model with n_profiles = 9 and model = 2

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 37/54

#> Processed model with n_profiles = 9 and model = 3

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 38/54

#> Processed model with n_profiles = 9 and model = 4

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 39/54

#> Processed model with n_profiles = 9 and model = 5

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 40/54

#> Processed model with n_profiles = 10 and model = 1

#> Result: BIC = 661.517

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 41/54

#> Processed model with n_profiles = 10 and model = 2

#> Result: LL not replicated

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 42/54

#> Processed model with n_profiles = 10 and model = 3

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 43/54

#> Processed model with n_profiles = 10 and model = 4

#> Result: Convergence problem

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 

#> Model 44/54

#> Processed model with n_profiles = 10 and model = 5

#> Result: Convergence problem

#>   n_profiles  model_1           model_2             model_3             model_4
#> 1          2 1042.968           688.097             857.551             672.333
#> 2          3  813.042           632.963             743.997             618.349
#> 3          4  735.479           591.406             705.067             607.269
#> 4          5  694.389           595.889             700.902             638.968
#> 5          6  672.595           598.972             704.198              638.26
#> 6          7  669.482           610.265             704.525             680.795
#> 7          8  661.074 LL not replicated   LL not replicated   LL not replicated
#> 8          9  661.624 LL not replicated   LL not replicated Convergence problem
#> 9         10  661.517 LL not replicated Convergence problem Convergence problem
#>               model_5
#> 1             574.018
#> 2             580.839
#> 3   LL not replicated
#> 4   LL not replicated
#> 5   LL not replicated
#> 6 Convergence problem
#> 7 Convergence problem
#> 8 Convergence problem
#> 9 Convergence problem