Create profiles for a specific mclust model

Create profiles for a specific mclust model

create_profiles_lpa(df, ..., n_profiles, model = 1, to_return = "tibble",
  center_raw_data = FALSE, scale_raw_data = FALSE,
  return_posterior_probs = TRUE, return_orig_df = FALSE,
  prior_control = FALSE)

Arguments

df
data.frame with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles
the number of profiles (or mixture components) to be estimated

model
the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to 0); 2 (varying means, equal variances and covariances; 3 (varying means and variances, covariances fixed to 0), 4 (varying means and covariances, equal variances; can only be specified in Mplus); and 5 (varying means, variances, and covariances), in order least to most freely-estimated; see the introductory vignette for more information

to_return
character string for either "tibble" or "mclust" if "tibble" is selected, then data with a column for profiles is returned; if "mclust" is selected, then output of class mclust is returned

center_raw_data
logical for whether to center (M = 1) the raw data (before clustering); defaults to FALSE

scale_raw_data
logical for whether to scale (SD = 1) the raw data (before clustering); defaults to FALSE

return_posterior_probs
TRUE or FALSE (only applicable if to_return == "tibble"); whether to include posterior probabilities in addition to the posterior profile classification; defaults to TRUE

return_orig_df
TRUE or FALSE (if TRUE, then the entire data.frame is returned; if FALSE, then only the variables used in the model are returned)

prior_control
whether to include a regularizing prior; defaults to false

Value

either a tibble or a ggplot2 plot of the BIC values for the explored models

Details

Creates profiles (or estimates of the mixture components) for a specific mclust model in terms of the specific number of mixture components and the structure of the residual covariance matrix

Examples

d <- pisaUSA15::pisaUSA15

#> Error: 'pisaUSA15' is not an exported object from 'namespace:pisaUSA15'

d <- dplyr::sample_n(d, 200)

#> Error: `size` must be less or equal than 9 (size of data), set `replace` = TRUE to use sampling with replacement

m3 <- create_profiles_lpa(d,
                          broad_interest, enjoyment, self_efficacy,
                          n_profiles = 3, to_return="tibble")

#> Error in .f(.x[[i]], ...): object 'broad_interest' not found