Create models for a specific mclust model

Create models for a specific mclust model

create_profiles_mplus(df, ..., n_profiles, the_title = "test",
  data_filename = "d.dat", script_filename = "t.inp",
  output_filename = "t.out", model = 1)

Arguments

df
data.frame with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles
the number of profiles (or mixture components) to be estimated

the_title
title of the model; defaults to test

data_filename
name of data file to prepare; defaults to d.dat

script_filename
name of script to prepare; defaults to t.inp

output_filename
name of the output; defaults to t.out

model
the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to 0); 2 (varying means, equal variances and covariances; 3 (varying means and variances, covariances fixed to 0), 4 (varying means and covariances, equal variances; can only be specified in Mplus); and 5 (varying means, variances, and covariances), in order least to most freely-estimated; see the introductory vignette for more information

Value

either a tibble or a ggplot2 plot of the BIC values for the explored models

Details

Creates an mplus model (.inp) and associated data file (.dat)

Examples

d <- tbl_df(iris[, -5])
m1 <- create_profiles_mplus(d,
                            Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                            n_profiles = 2,
                            model = 1)

#> TITLE: Your title goes here
#> DATA: FILE = "d.dat";
#> VARIABLE: 
#> NAMES = Sepal_Length Sepal_Width Petal_Length Petal_Width; 
#> MISSING=.;
#> 
#> Running model: t.inp 
#> System command: cd "/Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference" && "/Applications/Mplus/mplus" "t.inp" 
#> Reading model:  /Users/joshuarosenberg/Dropbox/1_Research/tidymixmod/docs/reference/t.out 
#>         LL      BIC Entropy
#> 1 -488.915 1042.968   0.991