Estimate latent profiles using Mplus

Estimates latent profiles (finite mixture models) using the commercial program Mplus, through the R-interface of MplusAutomation.

estimate_profiles_mplus2(
  df,
  n_profiles,
  model_numbers,
  select_vars,
  ...,
  keepfiles = FALSE
)

Arguments

df

data.frame with two or more columns with continuous variables

n_profiles

Numeric vector. The number of profiles (or mixture components) to be estimated. Each number in the vector corresponds to an analysis with that many mixture components.

model_numbers

Numeric vector. Numbers of the models to be estimated. See estimate_profiles for a description of the models available in tidyLPA.

select_vars

Character. Optional vector of variable names in df, to be used for model estimation. Defaults to NULL, which means all variables in df are used.

...

Parameters passed directly to mplusModeler. See the documentation of mplusModeler.

keepfiles

Logical. Whether to retain the files created by mplusModeler (e.g., for future reference, or to manually edit them).

Value

An object of class 'tidyLPA' and 'list'

Author

Caspar J. van Lissa