Extract log-likelihoods from models fit with estimate_profiles_mplus()

Extract log-likelihoods from models fit with estimate_profiles_mplus()

extract_LL_mplus(output_filename = "i.out")

Arguments

output_filename
name of output_filename from estimate_profiles_mplus()

Value

a tibble or a ggplot2 plot of the BIC values for the explored modelswith the log-likelihood, random start seed, and the number of the iteration

Details

Extract log-likelihoods associated with solutions from random starts from estimate_profiles_mplus(). Note that return_tmp_files = FALSE must be added to estimate_profiles_mplus() for this function to work.

Examples

not_run({
  m1 <- estimate_profiles_mplus(iris,
                              Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                              n_profiles = 2,
                              model = 1,
                              remove_tmp_files = FALSE)
  extract_LL_mplus(m1)
})