student questionnaire data with four variables from the 2015 PISA for students in the United States

student questionnaire data with four variables from the 2015 PISA for students in the United States

pisaUSA15

Format

Data frame with columns #'

CNTSTUID

international student ID

SCHID

international school ID

...

Source

http://www.oecd.org/pisa/data/