Select best model using analytic hyrarchy process

Integrates information from several fit indices, and selects the best model.

AHP(fitindices, relative_importance = c(AIC = 0.2323, AWE = 0.1129, BIC =
  0.2525, CLC = 0.0922, KIC = 0.3101))

Arguments

fitindices
A matrix or data.frame of fit indices, with colnames corresponding to the indices named in relative_importance.

relative_importance
A named numeric vector. Names should correspond to columns in fitindices, and values represent the relative weight assigned to the corresponding fit index. The default value corresponds to the fit indices and weights assigned by Akogul and Erisoglu. To assign uniform weights (i.e., each index is weighted equally), assign an equal value to all.

Value

Numeric.

Details

Many fit indices are available for model selection. Following the procedure developed by Akogul and Erisoglu (2017), this function integrates information from several fit indices, and selects the best model, using Saaty's (1990) Analytic Hierarchy Process (AHP). Conceptuall, the process consists of the following steps:

For each fit index, calculate the amount of support provided for each model, relative to the other models.

From these comparisons, obtain a "priority vector" of the amount of support for each model.

Compute a weighted average of the priority vectors for all fit indeces, with weights based on a simulation study examining each fit index' ability to recover the correct number of clusters (Akogul & Erisoglu, 2016).

Select the model with the highest weighted average priority.

Examples

iris[,1:4] %>%
  estimate_profiles(1:4) %>%
  get_fit() %>%
  AHP()

#> Warning: 
#> One or more analyses resulted in warnings! Examine these analyses carefully: model_1_class_3, model_1_class_4

#> [1] 3