Create models for a specific mclust model

Create models for a specific mclust model

create_profiles_mplus(df, ..., n_profiles, the_title = "test",
  data_filename = "d.dat", script_filename = "i.inp",
  output_filename = "i.out", savedata_filename = "d-mod.dat", model = 1,
  start_iterations = c(500, 50), m_iterations = 500, st_iterations = 50,
  convergence_criterion = 1e-07, remove_tmp_files = TRUE,
  print_input_file = FALSE, return_save_data = FALSE)

Arguments

df
data.frame with two or more columns with continuous variables

...
unquoted variable names separated by commas

n_profiles
the number of profiles (or mixture components) to be estimated

the_title
title of the model; defaults to test

data_filename
name of data file to prepare; defaults to d.dat

script_filename
name of script to prepare; defaults to i.inp

output_filename
name of the output; defaults to o.out

model
the mclust model to explore: 1 (varying means, equal variances, and residual covariances fixed to 0); 2 (varying means, equal variances and covariances; 3 (varying means and variances, covariances fixed to 0), 4 (varying means and covariances, equal variances; can only be specified in Mplus); 5 (varying means, equal variances, varying covariances); and 6 (varying means, variances, and covariances), in order least to most freely-estimated; see the introductory vignette for more information

start_iterations
the number of start iterations; defaults to c(500, 50)

m_iterations
number of m-step iterations; defaults to 500

convergence_criterion
the convergence criterion; defaults to .0000001

remove_tmp_files
whether to remove data, script, and output files; defaults to TRUE

print_input_file
whether to print the input file to the console

return_save_data
whether to return the save data (with the original data and the posterior probabiltiies for the classes and the class assignment) as a data.frame along with the MPlus output

s_iterations
the number of initial stage iterations; defaults to 50

Value

either a tibble or a ggplot2 plot of the BIC values for the explored models

Details

Creates an mplus model (.inp) and associated data file (.dat)

Examples

m1 <- create_profiles_mplus(iris,
                            Sepal.Length, Sepal.Width, Petal.Length, Petal.Width,
                            n_profiles = 2,
                            model = 1)
# models described here:
# https://jrosen48.github.io/tidyLPA/articles/Introduction_to_tidyLPA.html