Estimate latent profiles

Estimates latent profiles (finite mixture models) using the open source package mclust, or the commercial program Mplus (using the R-interface of MplusAutomation).

estimate_profiles(df, n_profiles, models = NULL, variances = "equal",
  covariances = "zero", package = "mclust", ...)

Arguments

df
data.frame of numeric data; continuous indicators are required for mixture modeling.

n_profiles
Integer vector of the number of profiles (or mixture components) to be estimated.

models
Integer vector. Set to NULL by default, and models are constructed from the variances and covariances arguments. See Details for the six models available in tidyLPA.

variances
Character vector. Specifies which variance components to estimate. Defaults to "equal" (constrain variances across profiles); the other option is "varying" (estimate variances freely across profiles). Each element of this vector refers to one of the models you wish to run.

covariances
Character vector. Specifies which covariance components to estimate. Defaults to "zero" (do not estimate covariances; this corresponds to an assumption of conditional independence of the indicators); other options are "equal" (estimate covariances between items, constrained across profiles), and "varying" (free covariances across profiles).

package
Character. Which package to use; 'mclust' or 'MplusAutomation' (requires Mplus to be installed). Default: 'mclust'.

...
Additional arguments are passed to the estimating function; i.e., Mclust, or mplusModeler.

Value

A list of class 'tidyLPA'.

Details

Six models are currently available in tidyLPA, corresponding to the most common requirements. These are:

Equal variances and covariances fixed to 0

Varying variances and covariances fixed to 0

Equal variances and equal covariances

Varying variances and equal covariances

Equal variances and varying covariances

Varying variances and varying covariances

Two interfaces are available to estimate these models; specify their numbers in the models argument (e.g., models = 1, or models = c(1, 2, 3)), or specify the variances/covariances to be estimated (e.g.,: variances = c("equal", "varying"), covariances = c("zero", "equal")).

Examples

iris_sample <- iris[c(1:4, 51:54, 101:104), ] # to make example run more quickly

# Example 1:
iris_sample %>%
  subset(select = c("Sepal.Length", "Sepal.Width",
    "Petal.Length")) %>%
  estimate_profiles(3)

#> tidyLPA analysis using mclust: 
#> 
#>  Model Classes AIC    BIC    Entropy prob_min prob_max n_min n_max BLRT_p
#>  1     3       62.153 68.942 0.998   0.999    1.000    0.167 0.500 0.020 

# Example 2:
iris %>%
  subset(select = c("Sepal.Length", "Sepal.Width",
    "Petal.Length")) %>%
  estimate_profiles(n_profiles = 1:4, models = 1:3)

#> The 'variances'/'covariances' arguments were ignored in favor of the 'models' argument.

#> tidyLPA analysis using mclust: 
#> 
#>  Model Classes AIC      BIC      Entropy prob_min prob_max n_min n_max BLRT_p
#>  1     1       1150.806 1168.870 1.000   1.000    1.000    1.000 1.000       
#>  1     2       885.638  915.744  0.985   0.995    0.995    0.340 0.660 0.010 
#>  1     3       765.160  807.309  0.925   0.943    0.999    0.267 0.400 0.010 
#>  1     4       758.711  812.902  0.866   0.700    0.961    0.093 0.393 0.020 
#>  2     1       1150.806 1168.870 1.000   1.000    1.000    1.000 1.000       
#>  2     2       770.958  810.096  1.000   1.000    1.000    0.333 0.667 0.010 
#>  2     3       702.546  762.758  0.887   0.907    1.000    0.313 0.353 0.010 
#>  2     4       695.509  776.796  0.876   0.907    0.949    0.093 0.353 0.050 
#>  3     1       851.331  878.427  1.000   1.000    1.000    1.000 1.000       
#>  3     2       693.264  732.402  1.000   1.000    1.000    0.333 0.667 0.010 
#>  3     3       650.144  701.325  0.917   0.944    1.000    0.327 0.340 0.010 
#>  3     4       646.806  710.029  0.896   0.706    0.967    0.047 0.340 0.040 

# Example 3:
iris_subset %>%
  subset(select = c("Sepal.Length", "Sepal.Width",
    "Petal.Length")) %>%
  estimate_profiles(n_profiles = 1:4, variances = c("equal", "varying"),
                    covariances = c("zero", "zero"))

#> Error in eval(lhs, parent, parent): object 'iris_subset' not found