Simulated identity data

This simulated dataset, based on Crochetti et al., 2014, contains five annual assessments of adolescents' mean scores on the commitment, exploration (in depth), and reconsideration subscales of the Utrecht-Management of Identity Commitments Scale (Crocetti et al., 2008). The scores reported here reflect the educational identity subscales of this instrument. The first measurement wave occurred when adolescents were, on average, 14 years old, and the last one when they were 18 years old.

data(id_edu)

Format

A data frame with 443 rows and 16 variables.

Details

com1	numeric	Mean score of educational commitment in wave 1

exp1	numeric	Mean score of educational exploration in wave 1

rec1	numeric	Mean score of educational reconsideration in wave 1

com2	numeric	Mean score of educational commitment in wave 2

exp2	numeric	Mean score of educational exploration in wave 2

rec2	numeric	Mean score of educational reconsideration in wave 2

com3	numeric	Mean score of educational commitment in wave 3

exp3	numeric	Mean score of educational exploration in wave 3

rec3	numeric	Mean score of educational reconsideration in wave 3

com4	numeric	Mean score of educational commitment in wave 4

exp4	numeric	Mean score of educational exploration in wave 4

rec4	numeric	Mean score of educational reconsideration in wave 4

com5	numeric	Mean score of educational commitment in wave 5

exp5	numeric	Mean score of educational exploration in wave 5

rec5	numeric	Mean score of educational reconsideration in wave 5

sex	factor	Adolescent sex; M = male, F = female.

References

Crocetti, E., Klimstra, T. A., Hale, W. W., Koot, H. M., & Meeus, W. (2013). Impact of early adolescent externalizing problem behaviors on identity development in middle to late adolescence: A prospective 7-year longitudinal study. Journal of Youth and Adolescence, 42(11), 1745-1758. DOI:10.1007/s10964-013-9924-6